Navigation Fixes #538
