+++ frontend/app/(dashboard)/[city]/massnahmen/page.module.scss
@@ -1,13 +1,6 @@
-
 .container {
   max-width: 50rem;
   margin-bottom: 5rem;
-}
-
-.accordion {
-  & > * {
-    margin-bottom: 2rem;
-  }
 }
 
 .legende {
