+++ frontend/app/(dashboard)/[city]/page.tsx
@@ -6,7 +6,8 @@ import greenCity from "@/public/background-green-city.webp";
 import LocalGroup from "@/app/components/LocalGroup";
 import NavigationTile from "@/app/components/NavigationTile";
 import styles from "./page.module.scss";
-import { getCities, getTasks, getRecursiveStatusNumbers } from "@/lib/dataService";
+import { getCities, getTasks } from "@/lib/dataService";
+import {  getRecursiveStatusNumbers } from "@/lib/utils";
 import ImplementationIndicator from "@/app/components/ImplementationIndicator";
 import ChecklistIndicator from "@/app/components/ChecklistIndicator";
 import type { CheckItem, Chart } from "@/types";
