+++ frontend/app/components/MeasureCardContent.tsx
@@ -1,9 +1,12 @@
-import React from "react";
+"use client";
+
+import React, { useEffect, useState } from "react";
 import { Accordion } from "react-bootstrap";
 import styles from "./styles/MeasureCard.module.scss";
 import LinkMeasureCard from "./LinkMeasureCard";
 import SecondaryMeasureCard from "./SecondaryMeasureCard";
 import { Task } from "@/types";
+import { useRouter, useSearchParams } from "next/navigation";
 
 interface MeasureCardContentProps {
   text: string;
@@ -13,12 +16,37 @@ interface MeasureCardContentProps {
   activeKey: string;
 }
 
-const MeasureCardContent: React.FC<MeasureCardContentProps> = ({ text, tasks, slugs, eventKey, activeKey="" }) => {
+const MeasureCardContent: React.FC<MeasureCardContentProps> = ({ text, tasks, slugs, eventKey}) => {
+  const searchParams = useSearchParams();
+  const router = useRouter();
+  const [activeKey, setActiveKey] = useState<string>("");
+
+  useEffect(() => {
+    const active = searchParams.get("activesub");
+    if (active) {
+      setActiveKey(active);
+    }
+  }, [searchParams]);
+
+  const handleSelect = (key: string | null) => {
+    const newKey = key || "";
+    setActiveKey(newKey);
+
+    const params = new URLSearchParams(searchParams.toString());
+
+    if (newKey) {
+      params.set("activesub", newKey);
+    } else {
+      params.delete("activesub");
+    }
+
+    router.replace(`?${params.toString()}`, { scroll: false });
+  };
 
   return (
     <div >
       {text} <a href={"./massnahmen/"+slugs}>Mehr lesen...</a>
-      <Accordion className={styles.contentaccordion} defaultActiveKey={activeKey}>
+      <Accordion className={styles.contentaccordion} activeKey={activeKey} onSelect={(key) => handleSelect(key as string)}>
         {tasks.map((task, i) => {
           if (task.numchild && task.children.length > 0) {
             return (
