+++ frontend/app/components/MeasuresAccordion.module.scss
@@ -1,0 +1,5 @@
+.accordion {
+  & > * {
+    margin-bottom: 2rem;
+  }
+}
