+++ frontend/lib/dataService.tsx
@@ -1,6 +1,5 @@
 import {cookies} from 'next/headers';
-import type { Task, StatusCount } from "@/types";
-import { ExecutionStatus } from "@/types/enums";
+
 
 const getCookie = async (name: string) => {
   return cookies().get(name)?.value ?? '';
@@ -38,37 +37,3 @@ export async function getTasks(id: string = "") {
 
 }
 
-export const getRecursiveStatusNumbers = (
-  tasks: Task[],
-): { complete: number; asPlanned: number; delayed: number; failed: number; unknown: number } => {
-  return tasks.reduce(
-    (statusCount, task) => {
-      if (task.numchild > 0) {
-        const childrenStatusNumbers = getRecursiveStatusNumbers(task.children);
-        Object.keys(childrenStatusNumbers).forEach((statusKey) => {
-          statusCount[statusKey as keyof StatusCount] += childrenStatusNumbers[statusKey as keyof StatusCount];
-        });
-      } else {
-        switch (task.execution_status) {
-          case ExecutionStatus.UNKNOWN:
-            statusCount.unknown++;
-            break;
-          case ExecutionStatus.AS_PLANNED:
-            statusCount.asPlanned++;
-            break;
-          case ExecutionStatus.COMPLETE:
-            statusCount.complete++;
-            break;
-          case ExecutionStatus.DELAYED:
-            statusCount.delayed++;
-            break;
-          case ExecutionStatus.FAILED:
-            statusCount.failed++;
-            break;
-        }
-      }
-      return statusCount;
-    },
-    { complete: 0, asPlanned: 0, delayed: 0, failed: 0, unknown: 0 } as StatusCount,
-  );
-};
